+++ src/stores/options.store.ts
@@ -1,3 +1,4 @@
+// 8KB item
 const { data: settings, promise } = useBrowserSyncStorage<settingsType>("settings", defaultSettings)
 export const useOptionsStore = defineStore("options", () => {
 	return {
@@ -5,6 +6,15 @@ export const useOptionsStore = defineStore("options", () => {
 	}
 })
 export const SettingsPromise = promise
+
+// Sync Data 8KB per item in firefox storage sync
+const { data: crunchyList, promise: crunchyListPromise } = useBrowserSyncStorage<CrunchyList>("crunchyList", [], false)
+export const crunchyListStore = defineStore("crunchyList", () => {
+	return {
+		crunchyList,
+	}
+})
+
 export const useFrontendStore = defineStore("frontend", () => {
 	const { isDark, toggleDark } = useTheme()
 	const currentLocale = useLocale()
@@ -17,6 +27,8 @@ export const useFrontendStore = defineStore("frontend", () => {
 export type BooleanObject = {
 	[key: string]: boolean
 }
+
+// 8KB item
 const { data: hideTitles, promise: hideTitlesPromise } = useBrowserSyncStorage<BooleanObject>("hideTitles", {}, false)
 export const useHideTitlesStore = defineStore("hideTitles", () => {
 	return {